requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
playwright==1.45.0
trafilatura==1.12.0
tiktoken==0.7.0
datasketch==1.6.5
langdetect==1.0.9
lancedb==0.9.0
umap-learn==0.5.6
scikit-learn==1.5.0
anthropic==0.34.0
# For Vertex AI support: pip install anthropic[vertex]  (adds google-auth)
# google-auth>=2.0.0
openai==1.40.0
httpx==0.27.0
python-dotenv==1.0.1
